+++ Stepic/en.lproj/Localizable.strings
@@ -259,3 +259,27 @@ RetentionNotificationYesterdayZero = "You didn't score any points for yesterday.
 RetentionNotificationYesterdayStreak = "You have been learning %@ in a row. Go back and keep learning!";
 RetentionNotificationWeekly = "You haven't done it for a long time. Go back and keep learning!";
 SocialSignupWithExistingEmail = "This email is already used. To log in using this social account you have to log in via email and password";
+AdaptiveAchievementFirstStep = "First steps";
+AdaptiveAchievementFirstStepDesc = "Complete tutorial";
+AdaptiveAchievementShare = "Sociable";
+AdaptiveAchievementShareDesc = "Share your achievement";
+AdaptiveAchievementExpDesc = "Gain %@ XP";
+AdaptiveAchievementExp1 = "Learner";
+AdaptiveAchievementExp2 = "Student";
+AdaptiveAchievementExp3 = "Experienced";
+AdaptiveAchievementExp4 = "Professional";
+AdaptiveAchievementExp5 = "Guru";
+AdaptiveAchievementStreakDesc = "Make %@ correct answers in a row";
+AdaptiveAchievementStreak1 = "Combo";
+AdaptiveAchievementStreak2 = "That's was easy";
+AdaptiveAchievementStreak3 = "Exactly!";
+AdaptiveAchievementStreak4 = "Clairvoyant";
+AdaptiveAchievementDaysStreakDesc = "Study for %@ days in a row";
+AdaptiveAchievementDaysStreak1 = "Enthusiast";
+AdaptiveAchievementDaysStreak2 = "Workweek";
+AdaptiveAchievementDaysStreak3 = "Addicted";
+AdaptiveAchievementDaysStreak4 = "Workaholic";
+AdaptiveAchievementDaysStreak5 = "Hero";
+AdaptiveAchievementLevelDesc = "Reach %@ level";
+AdaptiveAchievementLevel1 = "Expert";
+AdaptiveAchievementLevel2 = "Leader";
