+++ Stepic/ru.lproj/Localizable.strings
@@ -260,3 +260,27 @@ RetentionNotificationYesterdayZero = "За вчерашний день Вы ни
 RetentionNotificationYesterdayStreak = "Вы занимаетесь уже %@ подряд. Возвращайтесь и продолжайте учиться!";
 RetentionNotificationWeekly = "Вы давно не занимались. Возвращайтесь и продолжайте учиться!";
 SocialSignupWithExistingEmail = "Указанный email привязан к другому аккаунту. Чтобы входить через эту социальную сеть, необходимо войти, используя email и пароль";
+AdaptiveAchievementFirstStep = "Первые шаги";
+AdaptiveAchievementFirstStepDesc = "Пройти обучение";
+AdaptiveAchievementShare = "Общительный";
+AdaptiveAchievementShareDesc = "Поделиться любым своим достижением";
+AdaptiveAchievementExpDesc = "Получить %@ опыта";
+AdaptiveAchievementExp1 = "Ученик";
+AdaptiveAchievementExp2 = "Студент";
+AdaptiveAchievementExp3 = "Опытный";
+AdaptiveAchievementExp4 = "Профессионал";
+AdaptiveAchievementExp5 = "Гуру";
+AdaptiveAchievementStreakDesc = "Решить %@ заданий подряд с первой попытки";
+AdaptiveAchievementStreak1 = "Комбо";
+AdaptiveAchievementStreak2 = "Было просто";
+AdaptiveAchievementStreak3 = "В точку";
+AdaptiveAchievementStreak4 = "Ясновидящий";
+AdaptiveAchievementDaysStreakDesc = "Решать %@ дней подряд";
+AdaptiveAchievementDaysStreak1 = "Работяга";
+AdaptiveAchievementDaysStreak2 = "Трудовая неделя";
+AdaptiveAchievementDaysStreak3 = "Зависимый";
+AdaptiveAchievementDaysStreak4 = "Стахановец";
+AdaptiveAchievementDaysStreak5 = "Герой труда";
+AdaptiveAchievementLevelDesc = "Достигнуть %@ уровня";
+AdaptiveAchievementLevel1 = "Знаток";
+AdaptiveAchievementLevel2 = "Лидер";
