+++ StepicAdaptiveCourse/AchievementManager.swift
@@ -95,108 +95,3 @@ enum AchievementEvent {
     case days(value: Int)
     case share
 }
-
-extension AchievementManager {
-    static let shared = AchievementManager.createAndRegisterAchievements()
-
-    static func createAndRegisterAchievements() -> AchievementManager {
-        let mgr = AchievementManager()
-
-        let isOnboardingPassed = DefaultsStorageManager.shared.isRatingOnboardingFinished
-        let curRating = RatingManager.shared.rating
-        let curStreak = StatsManager.shared.maxStreak
-        let curLevel = RatingHelper.getLevel(for: curRating)
-
-        typealias ChallengeAchievementDescription = (slug: String, name: String, info: String, cover: UIImage, pre: ((Int, Int, Int) -> (Bool))?, migration: (() -> Int)?, event: String)
-        typealias ProgressAchievementDescription = (slug: String, name: String, info: String, cover: UIImage, maxValue: Int, pre: ((Int, Int, Int) -> (Bool))?, value: ((Int, Int, Int) -> Int)?, migration: (() -> Int)?, event: String)
-
-        let challengeAchievements: [ChallengeAchievementDescription] = [
-            (slug: "onboarding", name: "Первые шаги", info: "Пройти обучение", cover: #imageLiteral(resourceName: "onboarding"), pre: nil, migration: { return isOnboardingPassed ? 1 : 0 }, event: AchievementEvent.events.onboarding),
-            (slug: "share", name: "Общительный", info: "Поделиться любым своим достижением", cover: #imageLiteral(resourceName: "ashare"), pre: nil, migration: nil, event: AchievementEvent.events.share)
-        ]
-
-        func createExpAchievement(name: String, exp: Int, cover: UIImage) -> ProgressAchievementDescription {
-            return (slug: "exp\(exp)",
-                    name: "\(name)",
-                    info: "Получить \(exp) опыта",
-                    cover: cover,
-                    maxValue: exp,
-                    pre: nil,
-                    value: { cur, _, value in return cur + value },
-                    migration: { return curRating },
-                    event: AchievementEvent.events.exp)
-        }
-
-        func createStreakAchievement(name: String, streak: Int, cover: UIImage) -> ProgressAchievementDescription {
-            return (slug: "streak\(streak)",
-                name: "\(name)",
-                info: "Решить \(streak) заданий подряд с первой попытки",
-                cover: cover,
-                maxValue: streak,
-                pre: { cur, _, value in return value > cur },
-                value: { cur, _, value in return max(cur, value) },
-                migration: { return curStreak },
-                event: AchievementEvent.events.streak)
-        }
-
-        func createDaysAchievement(name: String, streak: Int, cover: UIImage) -> ProgressAchievementDescription {
-            return (slug: "days\(streak)",
-                name: "\(name)",
-                info: "Решать \(streak) дней подряд",
-                cover: cover,
-                maxValue: streak,
-                pre: { cur, _, value in return value > cur },
-                value: { cur, _, value in return max(cur, value) },
-                migration: { return 0 },
-                event: AchievementEvent.events.days)
-        }
-
-        func createLevelAchievement(name: String, level: Int, cover: UIImage) -> ProgressAchievementDescription {
-            return (slug: "level\(level)",
-                name: "\(name)",
-                info: "Достигнуть \(level) уровня",
-                cover: cover,
-                maxValue: level,
-                pre: { cur, _, value in return value > cur },
-                value: { cur, _, value in return max(cur, value) },
-                migration: { return curLevel },
-                event: AchievementEvent.events.level)
-        }
-
-        let progressAchievements: [ProgressAchievementDescription] = [
-            createExpAchievement(name: "Ученик", exp: 10, cover: #imageLiteral(resourceName: "exp1")),
-            createExpAchievement(name: "Студент", exp: 100, cover: #imageLiteral(resourceName: "exp2")),
-            createExpAchievement(name: "Опытный", exp: 500, cover: #imageLiteral(resourceName: "exp3")),
-            createExpAchievement(name: "Профессионал", exp: 5000, cover: #imageLiteral(resourceName: "exp4")),
-            createExpAchievement(name: "Гуру", exp: 10000, cover: #imageLiteral(resourceName: "exp5")),
-            createStreakAchievement(name: "Комбо", streak: 5, cover: #imageLiteral(resourceName: "streak1")),
-            createStreakAchievement(name: "Было просто", streak: 15, cover: #imageLiteral(resourceName: "streak2")),
-            createStreakAchievement(name: "В точку", streak: 30, cover: #imageLiteral(resourceName: "streak3")),
-            createStreakAchievement(name: "Ясновидящий", streak: 60, cover: #imageLiteral(resourceName: "streak4")),
-            createDaysAchievement(name: "Работяга", streak: 2, cover: #imageLiteral(resourceName: "days1")),
-            createDaysAchievement(name: "Трудовая неделя", streak: 5, cover: #imageLiteral(resourceName: "days2")),
-            createDaysAchievement(name: "Зависимый", streak: 10, cover: #imageLiteral(resourceName: "days3")),
-            createDaysAchievement(name: "Стахановец", streak: 14, cover: #imageLiteral(resourceName: "days4")),
-            createDaysAchievement(name: "Герой труда", streak: 30, cover: #imageLiteral(resourceName: "days5")),
-            createLevelAchievement(name: "Знаток", level: 5, cover: #imageLiteral(resourceName: "level1")),
-            createLevelAchievement(name: "Лидер", level: 10, cover: #imageLiteral(resourceName: "level2"))
-        ]
-
-        for achievementDescription in challengeAchievements {
-            let achievement = ChallengeAchievement(slug: achievementDescription.slug, name: achievementDescription.name, info: achievementDescription.info, cover: achievementDescription.cover, migration: achievementDescription.migration)
-            mgr.addSubscriber(for: achievementDescription.event, observer: achievement)
-            mgr.storedAchievements.append(achievement)
-        }
-
-        for achievementDescription in progressAchievements {
-            let achievement = ProgressAchievement(slug: achievementDescription.slug, name: achievementDescription.name, info: achievementDescription.info, cover: achievementDescription.cover, maxProgressValue: achievementDescription.maxValue, migration: achievementDescription.migration)
-            achievement.preConditions = achievementDescription.pre
-            achievement.migration = achievementDescription.migration
-            achievement.value = achievementDescription.value
-            mgr.addSubscriber(for: achievementDescription.event, observer: achievement)
-            mgr.storedAchievements.append(achievement)
-        }
-
-        return mgr
-    }
-}
